Notice for everyone who wants to compare fonts:
Go to view the SVN repository, like you normally do. Go to the folder that you get the sibelius files from, and there should be a sub-folder named "Fonts". Go into that folder, and load up the file called "How To Install Fonts.txt" that you will find there. It'll tell you the rest of the story.