Added an "_Lyrics" folder to put auto-generated lyrics in.
Just in case you're confused -- yes, there is already a "Lyrics" folder. I have an automatic script that extracts lyrics from the psalms, and puts them in a text file. However, it doesn't do a perfect job of it. So I want to keep the automatic output separate from the other Lyrics, which may well have been (or will be at some point) touched up and edited by a human being.

I'll be putting a new set of auto-generated lyrics in here at some stage soon.